app/requirements.txt
langchain
langchain-community
langchain-openai
langchain-chroma
openai
chromadb
python-dotenv
fastapi
uvicorn
pypdf